+++ src/main/java/net/thedragonteam/armorplus/ArmorPlus.java
@@ -44,11 +44,12 @@ public class ArmorPlus {
     public static final String LIB_VERSION = "1.10.2-1.0.3.0";
     public static final String MANTLE_VERSION = "1.10.2-1.0.0";
     public static final String TCONSTRUCT_VERSION = "1.10.2-2.5.6";
+    public static final String BAUBLES_VERSION = "1.3.0";
     public static final String MODID = "armorplus";
     public static final String MODNAME = "ArmorPlus";
     public static final String UPDATE_JSON = "http://fdn.redstone.tech/TheDragonTeam/armorplus/update.json";
     public static final boolean CAN_BE_DEACTIVATED = false;
-    public static final String DEPEND = "required-after:thedragonlib@[" + ArmorPlus.LIB_VERSION + ",);" + "after:tesla@[" + ArmorPlus.TESLA_VERSION + ",);" + "after:mantle@[" + ArmorPlus.MANTLE_VERSION + ",);" + "after:tconstruct@[" + ArmorPlus.TCONSTRUCT_VERSION + ",);";
+    public static final String DEPEND = "required-after:thedragonlib@[" + ArmorPlus.LIB_VERSION + ",);" + "after:tesla@[" + ArmorPlus.TESLA_VERSION + ",);" + "after:mantle@[" + ArmorPlus.MANTLE_VERSION + ",);" + "after:tconstruct@[" + ArmorPlus.TCONSTRUCT_VERSION + ",);" + "after:Baubles@[" + ArmorPlus.BAUBLES_VERSION + ",);";
     public static final String GUIFACTORY = "net.thedragonteam.armorplus.client.gui.ConfigGuiFactory";
     public static final String CLIENTPROXY = "net.thedragonteam.armorplus.proxy.ClientProxy";
     public static final String SERVERPROXY = "net.thedragonteam.armorplus.proxy.ServerProxy";