+++ src/main/java/net/thedragonteam/armorplus/registry/ModCompatibility.java
@@ -6,6 +6,7 @@ package net.thedragonteam.armorplus.registry;
 
 import net.minecraftforge.fml.common.Loader;
 import net.thedragonteam.armorplus.compat.ICompatibility;
+import net.thedragonteam.armorplus.compat.baubles.CompatibilityBaubles;
 import net.thedragonteam.armorplus.compat.jei.CompatibilityJustEnoughItems;
 
 import java.util.ArrayList;
@@ -20,6 +21,7 @@ public class ModCompatibility {
 
     public static void registerModCompat() {
         compatibilities.add(new CompatibilityJustEnoughItems());
+        compatibilities.add(new CompatibilityBaubles());
     }
 
     public static void loadCompat(ICompatibility.InitializationPhase phase) {